📑 :: 프로젝트 구조 세팅
