+++ src/main/java/com/example/soopgwan/domain/achive/persistence/Achieve.java
@@ -1,5 +1,6 @@
-package com.example.soopgwan.domain.habit.domain;
+package com.example.soopgwan.domain.achive.persistence;
 
+import com.example.soopgwan.domain.user.persistence.User;
 import com.example.soopgwan.global.entity.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -9,22 +10,33 @@ import lombok.experimental.SuperBuilder;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @SuperBuilder
-@Table(name = "tbl_day")
+@Table(name = "tbl_achieve")
 @Entity
-public class Day extends BaseTimeEntity {
+public class Achieve extends BaseTimeEntity {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(columnDefinition = "VARCHAR(200)", nullable = false)
-    private String eval;
+    private String unlockCondition;
+
+    @Column(columnDefinition = "VARCHAR(10)", nullable = false)
+    private String title;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
 }
