+++ src/main/java/com/example/soopgwan/domain/achive/persistence/AchieveSuccessId.java
@@ -1,4 +1,4 @@
-package com.example.soopgwan.domain.habit.domain;
+package com.example.soopgwan.domain.achive.persistence;
 
 import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;
@@ -10,9 +10,9 @@ import java.io.Serializable;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @EqualsAndHashCode
-public class HabitSuccessId implements Serializable {
+public class AchieveSuccessId implements Serializable {
 
-    private Long id;
+    private Long user;
 
-    private Long habit;
+    private Long achieve;
 }
