--- src/main/java/com/example/soopgwan/domain/user/domain/User.java
@@ -1,44 +1,0 @@
-package com.example.soopgwan.domain.user.domain;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.validator.constraints.Length;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@Builder
-@Table(name = "tbl_user")
-@Entity
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(columnDefinition = "VARCHAR(24)", nullable = false)
-    private String accountId;
-
-    @Column(columnDefinition = "CHAR(60)", nullable = false)
-    private String password;
-
-    @Column(columnDefinition = "VARCHAR(10)", nullable = false)
-    private String name;
-
-    @Column(columnDefinition = "CHAR(11)", nullable = false)
-    private String phoneNumber;
-
-    @Column(nullable = false)
-    private Integer level;
-}
-
